%%    KATforDCEMRI: a Kinetic Analysis Tool for DCE-MRI
%%    Copyright 2014 Genentech, Inc.
%%
%%    For questions or comments, please contact
%%    Gregory Z. Ferl, Ph.D.
%%    Genentech, Inc.
%%    Development Sciences
%%    1 DNA Way, Mail stop 463A
%%    South San Francisco, CA, United States of America
%%    E-mail: ferl.gregory@gene.com

\name{KATforDCEMRI-package}
\Rdversion{1.1}
\alias{KATforDCEMRI-package}
\alias{KATforDCEMRI}
\docType{package}
\title{
R package for kinetic analysis of DCE-MRI data.
}
\description{
Package for kinetic analysis of longitudinal voxel-wise Dynamic Contrast Enhanced MRI data. Includes tools for visualization and exploration of voxel-wise parametric maps.

  Thanks to Ruediger Port (Genentech), David Clayton (Genentech), Lu Xu (OncoMed),
  Frank-Peter Theil (UCB) and Sheye Aliu (UCSF) for extensive testing and/or
  helpful technical suggestions and Michael Lawrence (Genentech) for providing
  general advice on R package development. Thanks to Alex de Crespigny
  (Genentech Clinical Imaging) and Paul Fielder (Genentech PKPD) for
  supporting development of KATforDCEMRI. Special Thanks to Gregoire Pau (Genentech)
  for code optimization and Daniel Barboriak (Duke University) for
  providing DCE-MRI data for package testing.
}
\details{
\tabular{ll}{
Package: \tab KATforDCEMRI\cr
Type: \tab Package\cr
Version: \tab 0.740\cr
Date: \tab 2014-02-04\cr
License: \tab LGPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
  Gregory Z. Ferl <ferl.gregory@gene.com> and Georges Hankov <hankovg@gene.com>.
}
\keyword{ package }
